Separated functions and functors for Polygons from regular utils
